Moved RegExPatternExtended to osw.utils.strings
Some example RegExPatternExtended usage is given
