+++ scripts/migration/file_page_migration.py
@@ -7,13 +7,13 @@ from pathlib import Path
 from typing import Dict, List
 
 import mwclient
-from regex_pattern import REGEX_PATTERN_LIB, REGEX_PATTERN_LIST
 
 import osw.data.import_utility as iu
 import osw.wiki_tools as wt
 from osw.core import OSW
-from osw.data.mining import match_first_regex_pattern, test_regex_pattern
 from osw.model.entity import Label, WikiFile
+from osw.utils.regex_pattern import REGEX_PATTERN_LIB, REGEX_PATTERN_LIST
+from osw.utils.strings import match_first_regex_pattern, test_regex_pattern
 from osw.utils.util import parallelize
 from osw.wtsite import WtPage, WtSite
